+++ commerce/delivery/src/main/java/ru/yandex/practicum/commerce/delivery/service/DeliveryService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.commerce.delivery.service;
 import ru.yandex.practicum.commerce.interaction.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.commerce.interaction.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -15,5 +16,5 @@ public interface DeliveryService {
 
     void deliveryFailed(UUID orderId);
 
-    Double calculateDeliveryCost(OrderDto orderDto);
+    BigDecimal calculateDeliveryCost(OrderDto orderDto);
 }
